+++ src/effects/specialMonsters.js
@@ -24,7 +24,7 @@ export function baseMonsterFeatureEffect(document, label) {
     flags: {
       dae: {
         transfer: false,
-        stackable: "none",
+        stackable: "noneName",
       },
       ddbimporter: {
         disabled: false,