+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Effects: Fix Darkness macros for v11.
 - Character: School of Enchantment: Hypnotic Gaze incorrectly set limited uses.
+- Monster effects: MidiQoL `actionSave` effects. The parser will try and add overtime effects for saves where characters can use their action to shake off an effect such as the Deathlock Mastermind Grave Bolt.
 
 # 3.4.40